ai-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
openai==1.3.7
scikit-learn==1.3.2
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
